clock.java -> universe_settings
